pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zte-l13"
version = "0.6.1"
description = "ZTE L13"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: 実機デバイスが必要なテスト",
]
